server/requirements.txt
asgiref==3.4.1
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
sqlparse==0.4.2
Werkzeug==2.0.2
